example/requirements.txt
# Your app requirements.
-r ../requirements_test.txt

# Your app in editable mode.
-e ../


Django==3.1.7
channels_redis==3.2.0
